Fleeing done

+++ Application/Source/Utils/Heuristic.h
@@ -2,17 +2,17 @@
 
 namespace Heuristic
 {
-    inline unsigned int Manhattan(const glm::vec2& v1, const glm::vec2& v2,
+    inline unsigned int Manhattan(const glm::ivec2& v1, const glm::ivec2& v2,
         int weight)
     {
-        glm::vec2 delta = v2 - v1;
+        glm::ivec2 delta = v2 - v1;
         return static_cast<unsigned int>(weight * (delta.x + delta.y));
     }
 
-    inline unsigned int Euclidean(const glm::vec2& v1, const glm::vec2& v2,
+    inline unsigned int Euclidean(const glm::ivec2& v1, const glm::ivec2& v2,
         int weight)
     {
-        glm::vec2 delta = v2 - v1;
+        glm::ivec2 delta = v2 - v1;
         return static_cast<unsigned int>(weight * sqrt(
             (delta.x * delta.x) + (delta.y * delta.y)
         ));
